+++ server/component/style/login/LoginController.php
@@ -49,7 +49,11 @@ class LoginController extends BaseController
             }
         } else {
             if (isset($_POST['type']) && $_POST['type'] == 'login' && isset($_POST['email']) && isset($_POST['password'])) {
-                if ($model->check_login_credentials($_POST['email'], $_POST['password'])) {
+                $res = $model->check_login_credentials($_POST['email'], $_POST['password']);
+                if ($res === '2fa') {
+                    // redirect to 2fa page
+                    header('Location: ' . $model->get_link_url(SH_TWO_FACTOR_AUTHENTICATION));
+                } else if ($res) {
                     if (isset($_POST['mobile']) && $_POST['mobile']) {
                         // set device id for the user
                         $device_token = isset($_POST['device_token']) ? $_POST['device_token'] : 'web';
@@ -58,8 +62,9 @@ class LoginController extends BaseController
                     } else {
                         header('Location: ' . $model->get_target_url());
                     }
-                } else
+                } else {
                     $this->failed = true;
+                }
             }
         }
     }